IKASAN-2135 finalising job lock management for locks associated with JobInstance and ContextInstances.

+++ ikasaneip/job-orchestration/builder/src/main/java/org/ikasan/job/orchestration/builder/context/JobLockBuilder.java
@@ -13,7 +13,7 @@ public class JobLockBuilder {
 
     private String lockName;
     private Map<String, List<SchedulerJob>> schedulerJobs;
-    private long lockCount = 1;
+    private int lockCount = 1;
 
     public JobLockBuilder() {
     }
@@ -36,7 +36,7 @@ public class JobLockBuilder {
         return this;
     }
 
-    public JobLockBuilder withLockCount(Long lockCount) {
+    public JobLockBuilder withLockCount(Integer lockCount) {
         this.lockCount = lockCount;
         return this;
     }
